pyproject.toml
[project]
name = "ollama-benchmarking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "openai>=1.72.0",
    "python-dotenv>=1.1.0",
    "uvicorn[standard]>=0.34.0",
    "weave>=0.51.42",
]
